pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "process-control-causal-ml"
version = "0.1.0"
description = "Autonomous Process Control with Causal ML"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "pandas>=2.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "econml>=0.15",
    "dowhy>=0.11",
    "causal-learn>=0.1.3",
    "lingam>=1.8",
    "pgmpy>=0.1.25",
    "networkx>=3.2",
    "matplotlib>=3.8",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "loguru>=0.7",
    "great-expectations>=0.18",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "dash>=2.16",
    "dash-bootstrap-components>=1.6",
    "plotly>=5.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "ty>=0.0.1",
    "httpx>=0.27",
    "pip-audit>=2.7",
    "bandit>=1.7",
    "radon>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/process_control_causal_ml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
